[XLA:CPU] Graph visualization called only for executor owned by CpuExecutable
At the moment graph visualization is always called. This is undesired behaviour as it will lead to too many graph creations that aren't needed, and pollute the log with unnecessary information.

Future plans are to have subgraphs of thunks such as conditional, while and call as part of the main graph.

PiperOrigin-RevId: 756664077